# Sito Web/api/swagger-output.json
{
  "swagger": "2.0",
  "info": {
    "title": "My API",
    "description": "Description",
    "version": "1.0.0"
  },
  "host": "localhost:3100",
  "basePath": "/",
  "schemes": [
    "http"
  ],
  "paths": {
    "/api/get_json": {
      "post": {
        "tags": [
          "Principale"
        ],
        "summary": "Richieste a Spotify",
        "description": "",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "url": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
    "/playlist/{id}": {
      "get": {
        "tags": [
          "Playlist"
        ],
        "summary": "Carica e renderizza la pagina della playlist di Spotify",
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {}
      }
    },
    "/my_playlist/{id}": {
      "get": {
        "tags": [
          "Playlist"
        ],
        "summary": "Carica e renderizza la pagina della playlist interna di MongoDB",
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {}
      }
    },
    "/album/{id}": {
      "get": {
        "tags": [
          "Album"
        ],
        "summary": "Carica e renderizza la pagina dell'albu di Spotify",
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {}
      }
    },
    "/artist/{id}": {
      "get": {
        "tags": [
          "Artista"
        ],
        "summary": "Carica e renderizza la pagina dell'artista di Spotify",
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {}
      }
    },
...
    "/change_my_playlist/{id}": {
      "get": {
        "tags": [
          "Playlist"
        ],
        "summary": "Renderizza la pagina per la modifica delle mie playlist interne",
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {}
      }
    },
...
    "/api/register": {
      "put": {
        "tags": [
          "Autentificazione"
        ],
        "summary": "Registra l'utente dentro",
        "description": "",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "password": {
                  "example": "any"
                },
                "username": {
                  "example": "any"
                },
                "email": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/api/login": {
      "post": {
        "tags": [
          "Autentificazione"
        ],
        "summary": "Fase di controllo per il login",
        "description": "",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "password": {
                  "example": "any"
                },
                "username": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
...
  }
}